Added score rendering + bug fix
Score Rendering: The score is now displayed

Bug Fix: I think that my enemy initalization had an incorret bound for the x, so an enemy could spawn well off the screen. Every frame it would check if the enemy was off screen, and finding that it was, would move the enemy down 40 px ending the game in 15 frames.